recplace Unicode stars by glyphicons
Might as well use them since we have them. The support is likely
to be as good as or better than Unicode support.

There's a glyphicon-star-half included in the full Glyphicon set,
but not in the Halflings set that's provided with Bootstrap. I'm
thinking if we ever need to make a half star, we could probably
CSS up a solution. For example, here's one:
http://codepen.io/leogono/pen/pvaKNO