Switch on iCheck
Finally. And the result is better, visually.

I was unable to avoid a small CSS tweak, just to get enough space
before the checkboxes. Unsure whether to mention that in the
course or not. It's pretty typical of the CSS hacks you have to
make. But on the other hand, this is a JavaScript course.